refactor(console): extract the oidc endpoint fields into a shared component
