+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/EndpointsAndCredentials/index.module.scss
@@ -1,14 +1,5 @@
 @use '@/scss/underscore' as _;
 
-.fieldButton {
-  margin-top: _.unit(2);
-}
-
-.trailingIcon {
-  width: 16px;
-  height: 16px;
-}
-
 button.add {
   margin-top: _.unit(2);
 }
